# docs/swagger.yaml
basePath: /v1
definitions:
...
  swaggertypes.BillingURLData:
    properties:
      url:
        example: https://app.lemonsqueezy.com/checkout/...
        type: string
    type: object
  swaggertypes.BillingURLResponse:
    properties:
      data:
        $ref: '#/definitions/swaggertypes.BillingURLData'
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
  swaggertypes.CaptureInboxRequest:
    properties:
      title:
        example: Read Go spec
        type: string
    type: object
...
  swaggertypes.CreateProjectRequest:
    properties:
      areaId:
        example: area_01jw...
        type: string
      name:
        example: Launch blog
        type: string
    type: object
...
  swaggertypes.EmptyData:
    type: object
  swaggertypes.EmptyResponse:
    properties:
      data:
        $ref: '#/definitions/swaggertypes.EmptyData'
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
  swaggertypes.ErrorBody:
    properties:
      code:
        example: INVALID_INPUT
        type: string
      message:
        example: email is required
        type: string
    type: object
  swaggertypes.ErrorResponse:
    properties:
      data:
        type: object
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
...
  swaggertypes.ProjectData:
    properties:
      area_id:
        example: area_01jw...
        type: string
      created_at:
        type: string
      display_order:
        example: 0
        type: integer
      id:
        example: proj_01jw...
        type: string
      name:
        example: Launch blog
        type: string
      status:
        example: active
        type: string
      updated_at:
        type: string
      user_id:
        example: usr_01jw...
        type: string
    type: object
  swaggertypes.ProjectListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/swaggertypes.ProjectData'
        type: array
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
  swaggertypes.ProjectResponse:
    properties:
      data:
        $ref: '#/definitions/swaggertypes.ProjectData'
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
...
  swaggertypes.TaskData:
    properties:
      created_at:
        type: string
      due_date:
        example: "2026-04-01"
        type: string
      id:
        example: task_01jw...
        type: string
      notes:
        example: 'See PR #42'
        type: string
      project_id:
        example: proj_01jw...
        type: string
      scheduled_for:
        example: today
        type: string
      status:
        example: active
        type: string
      title:
        example: Write unit tests
        type: string
      updated_at:
        type: string
      user_id:
        example: usr_01jw...
        type: string
    type: object
  swaggertypes.TaskListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/swaggertypes.TaskData'
        type: array
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
  swaggertypes.TaskResponse:
    properties:
      data:
        $ref: '#/definitions/swaggertypes.TaskData'
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
...
  swaggertypes.UpdateProjectRequest:
    properties:
      areaId:
        example: area_01jw...
        type: string
      name:
        example: Launch blog v2
        type: string
    type: object
...
  swaggertypes.UserPlanData:
    properties:
      created_at:
        type: string
      current_period_end:
        type: string
      current_period_start:
        type: string
      id:
        example: plan_01jw...
        type: string
      lemon_squeezy_customer_id:
        type: string
      lemon_squeezy_subscription_id:
        type: string
      plan:
        example: free
        type: string
      status:
        example: active
        type: string
      updated_at:
        type: string
      user_id:
        example: usr_01jw...
        type: string
    type: object
  swaggertypes.UserPlanResponse:
    properties:
      data:
        $ref: '#/definitions/swaggertypes.UserPlanData'
      error:
        $ref: '#/definitions/swaggertypes.ErrorBody'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for the Nicoflow productivity platform
  title: Nicoflow API
  version: "1.0"
paths:
...
  /billing/checkout-url:
    get:
      description: Returns a Lemon Squeezy checkout URL to upgrade to PRO
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggertypes.BillingURLResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get checkout URL
      tags:
      - billing
  /billing/portal-url:
    get:
      description: Returns a Lemon Squeezy customer portal URL to manage subscription
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggertypes.BillingURLResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get billing portal URL
      tags:
      - billing
  /billing/webhook:
    post:
      consumes:
      - application/json
      description: Receives and processes Lemon Squeezy webhook events. Idempotent
        — duplicate events are ignored.
      parameters:
      - description: HMAC-SHA256 signature from Lemon Squeezy
        in: header
        name: X-Signature
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggertypes.EmptyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      summary: Lemon Squeezy webhook
      tags:
      - billing
...
  /inbox:
    get:
      description: Returns all unscheduled tasks (project_id IS NULL) for the authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggertypes.TaskListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List inbox items
      tags:
      - inbox
  /inbox/capture:
    post:
      consumes:
      - application/json
      description: Quickly captures a task title into the inbox (no project, no scheduling)
      parameters:
      - description: Task title
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/swaggertypes.CaptureInboxRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/swaggertypes.TaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Capture to inbox
      tags:
      - inbox
...
  /projects:
    get:
      description: Returns all projects belonging to the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggertypes.ProjectListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Creates a new project for the authenticated user, optionally under
        an area
      parameters:
      - description: Project data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/swaggertypes.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/swaggertypes.ProjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a project
      tags:
      - projects
  /projects/{id}:
    delete:
      description: Deletes a project owned by the authenticated user
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggertypes.EmptyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a project
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Updates a project owned by the authenticated user
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated project data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/swaggertypes.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggertypes.ProjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a project
      tags:
      - projects
...
  /user/plan:
    get:
      description: Returns the current subscription plan for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swaggertypes.UserPlanResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user plan
      tags:
      - user
  /users/me:
    delete:
      description: Soft-deletes the authenticated user and revokes all their refresh
        tokens
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swaggertypes.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete account
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"